Clean up commands
